# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meme Tracker API",
    "description": "High-performance Python FastAPI service and background aggregation engine that continuously discovers, curates, ranks, and serves the newest and trending memes from popular internet sources (Reddit and Know Your Meme).",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/memes/latest": {
      "get": {
        "tags": [
          "memes"
        ],
        "summary": "Get newest memes",
        "description": "Retrieve newest memes sorted by publication date descending with pagination and filtering.",
        "operationId": "get_latest_memes_api_v1_memes_latest_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page limit (1-100 items)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page limit (1-100 items)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset index",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset index"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source platform (reddit, knowyourmeme) or community (r/memes, dankmemes)",
              "title": "Source"
            },
            "description": "Filter by source platform (reddit, knowyourmeme) or community (r/memes, dankmemes)"
          },
          {
            "name": "nsfw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include NSFW content if true (default false)",
              "default": false,
              "title": "Nsfw"
            },
            "description": "Include NSFW content if true (default false)"
          },
          {
            "name": "time_window",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional time window filter (e.g., 1h, 6h, 24h, 7d, all)",
              "title": "Time Window"
            },
            "description": "Optional time window filter (e.g., 1h, 6h, 24h, 7d, all)"
          },
          {
            "name": "generation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional generational filter (gen_alpha, gen_z, millennial, gen_x)",
              "title": "Generation"
            },
            "description": "Optional generational filter (gen_alpha, gen_z, millennial, gen_x)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of newest memes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on query parameters"
          }
        }
      }
    },
    "/api/v1/memes/trending": {
      "get": {
        "tags": [
          "memes"
        ],
        "summary": "Get trending memes",
        "description": "Retrieve trending memes ranked by virality/velocity score with gravity decay, pagination, and filtering.",
        "operationId": "get_trending_memes_api_v1_memes_trending_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page limit (1-100 items)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page limit (1-100 items)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset index",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset index"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source platform (reddit, knowyourmeme) or community (r/memes, dankmemes)",
              "title": "Source"
            },
            "description": "Filter by source platform (reddit, knowyourmeme) or community (r/memes, dankmemes)"
          },
          {
            "name": "nsfw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include NSFW content if true (default false)",
              "default": false,
              "title": "Nsfw"
            },
            "description": "Include NSFW content if true (default false)"
          },
          {
            "name": "time_window",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional time window filter (e.g., 1h, 6h, 24h, 7d, all)",
              "title": "Time Window"
            },
            "description": "Optional time window filter (e.g., 1h, 6h, 24h, 7d, all)"
          },
          {
            "name": "generation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional generational filter (gen_alpha, gen_z, millennial, gen_x)",
              "title": "Generation"
            },
            "description": "Optional generational filter (gen_alpha, gen_z, millennial, gen_x)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of trending memes sorted by trending_score",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error on query parameters"
          }
        }
      }
    },
    "/api/v1/memes/random": {
      "get": {
        "tags": [
          "memes"
        ],
        "summary": "Get a random meme",
        "description": "Fetch a single pseudo-random meme matching optional source and NSFW filters.",
        "operationId": "get_random_meme_api_v1_memes_random_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional source filter (reddit, knowyourmeme, dankmemes, etc.)",
              "title": "Source"
            },
            "description": "Optional source filter (reddit, knowyourmeme, dankmemes, etc.)"
          },
          {
            "name": "nsfw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include NSFW content if true (default false)",
              "default": false,
              "title": "Nsfw"
            },
            "description": "Include NSFW content if true (default false)"
          },
          {
            "name": "generation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional generational filter (gen_alpha, gen_z, millennial, gen_x)",
              "title": "Generation"
            },
            "description": "Optional generational filter (gen_alpha, gen_z, millennial, gen_x)"
          }
        ],
        "responses": {
          "200": {
            "description": "Random meme payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meme"
                }
              }
            }
          },
          "404": {
            "description": "No memes found matching the specified criteria"
          },
          "422": {
            "description": "Validation error on query parameters"
          }
        }
      }
    },
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List ingestion sources",
        "description": "Retrieve status, cached item counts, latency, and error metrics for all ingestion sources.",
        "operationId": "get_sources_api_v1_sources_get",
        "responses": {
          "200": {
            "description": "List of tracked data sources and their operational status",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceStatus"
                  },
                  "type": "array",
                  "title": "Response Get Sources Api V1 Sources Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Service health check",
        "description": "Retrieve service operational status, uptime, cached meme counts, and healthy source counts.",
        "operationId": "get_api_health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Service health metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Service health check",
        "description": "System health check returning operational status, uptime, cached memes, and healthy source counts.",
        "operationId": "root_health_health_get",
        "responses": {
          "200": {
            "description": "Service health metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Root index",
        "description": "Service overview and documentation links, or web UI for browsers.",
        "operationId": "root_index__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/web": {
      "get": {
        "tags": [
          "memes"
        ],
        "summary": "Web Explorer & Documentation Portal",
        "description": "Interactive Meme Explorer and API documentation portal.",
        "operationId": "web_portal_web_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Overall service health: ok, degraded, or unhealthy",
            "default": "ok"
          },
          "app_name": {
            "type": "string",
            "title": "App Name",
            "description": "Service name",
            "default": "Meme Tracker API"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Application version",
            "default": "0.1.0"
          },
          "uptime_seconds": {
            "type": "number",
            "title": "Uptime Seconds",
            "description": "Seconds since service start",
            "default": 0.0
          },
          "total_memes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Memes",
            "description": "Total distinct memes currently in cache",
            "default": 0
          },
          "total_memes_cached": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Memes Cached",
            "description": "Alias for total_memes",
            "default": 0
          },
          "healthy_sources": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy Sources",
            "description": "Count of healthy sources",
            "default": 0
          },
          "total_sources": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Sources",
            "description": "Count of total sources",
            "default": 0
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "type": "array",
            "title": "Sources",
            "description": "Health status of all data sources"
          }
        },
        "type": "object",
        "title": "HealthResponse",
        "description": "Response payload for GET /health."
      },
      "MediaType": {
        "type": "string",
        "enum": [
          "image",
          "gif",
          "video",
          "link"
        ],
        "title": "MediaType",
        "description": "Supported media asset types."
      },
      "Meme": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique meme ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Meme title"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Direct media URL"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url",
            "description": "Direct media URL (alias for url)"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType",
            "description": "Media type",
            "default": "image"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source platform name"
          },
          "source_platform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourcePlatform"
              },
              {
                "type": "string"
              }
            ],
            "title": "Source Platform",
            "description": "Source platform enum"
          },
          "source_community": {
            "type": "string",
            "title": "Source Community",
            "description": "Source community / subreddit / feed"
          },
          "permalink": {
            "type": "string",
            "title": "Permalink",
            "description": "Permalink to original post"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "description": "Author username",
            "default": "unknown"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "description": "Engagement score / upvotes",
            "default": 0
          },
          "num_comments": {
            "type": "integer",
            "title": "Num Comments",
            "description": "Comment count",
            "default": 0
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "description": "Creation epoch timestamp in UTC"
          },
          "is_nsfw": {
            "type": "boolean",
            "title": "Is Nsfw",
            "description": "NSFW flag",
            "default": false
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Media domain",
            "default": ""
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash",
            "description": "Content hash",
            "default": ""
          },
          "trending_score": {
            "type": "number",
            "title": "Trending Score",
            "description": "Trending score",
            "default": 0.0
          },
          "generation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemeGeneration"
              },
              {
                "type": "string"
              }
            ],
            "title": "Generation",
            "description": "Generational alignment (gen_alpha, gen_z, millennial, gen_x)",
            "default": "gen_z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "media_url",
          "source",
          "source_platform",
          "source_community",
          "permalink",
          "created_at"
        ],
        "title": "Meme",
        "description": "Public API response schema representing a meme."
      },
      "MemeGeneration": {
        "type": "string",
        "enum": [
          "all",
          "gen_alpha",
          "gen_z",
          "millennial",
          "gen_x"
        ],
        "title": "MemeGeneration",
        "description": "Generational internet culture eras."
      },
      "NormalizedMeme": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Globally unique meme identifier (e.g. reddit_memes_1d8xyz or kym_Entry-57336)"
          },
          "raw_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Id",
            "description": "Original platform identifier"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Meme headline or title"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url",
            "description": "Direct URL to media file (image, gif, mp4)"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType",
            "description": "Type of media asset (image, gif, video, link)",
            "default": "image"
          },
          "source_platform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourcePlatform"
              },
              {
                "type": "string"
              }
            ],
            "title": "Source Platform",
            "description": "Originating platform (reddit or knowyourmeme)"
          },
          "source_community": {
            "type": "string",
            "title": "Source Community",
            "description": "Subreddit or KYM category (e.g. r/memes or confirmed)"
          },
          "permalink": {
            "type": "string",
            "title": "Permalink",
            "description": "Canonical link to original post/entry"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "description": "Author or creator username",
            "default": "unknown"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "description": "Upvotes or baseline engagement score",
            "default": 0
          },
          "num_comments": {
            "type": "integer",
            "title": "Num Comments",
            "description": "Comment count",
            "default": 0
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "description": "Unix epoch timestamp in seconds (UTC)"
          },
          "is_nsfw": {
            "type": "boolean",
            "title": "Is Nsfw",
            "description": "NSFW flag",
            "default": false
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Hosting domain of media asset",
            "default": ""
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash",
            "description": "Deterministic SHA-256 hash for deduplication"
          },
          "trending_score": {
            "type": "number",
            "title": "Trending Score",
            "description": "Calculated engagement velocity score",
            "default": 0.0
          },
          "generation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemeGeneration"
              },
              {
                "type": "string"
              }
            ],
            "title": "Generation",
            "description": "Generational alignment (gen_alpha, gen_z, millennial, gen_x)",
            "default": "gen_z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "media_url",
          "source_platform",
          "source_community",
          "permalink",
          "created_at"
        ],
        "title": "NormalizedMeme",
        "description": "Internal normalized meme data contract produced by ingestion engine."
      },
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Meme"
                },
                {
                  "$ref": "#/components/schemas/NormalizedMeme"
                }
              ]
            },
            "type": "array",
            "title": "Items",
            "description": "List of meme items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total matching meme count",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Page limit",
            "default": 20
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Page offset",
            "default": 0
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Flag indicating if more items exist",
            "default": false
          }
        },
        "type": "object",
        "title": "PaginatedResponse",
        "description": "Paginated collection of meme items."
      },
      "SourcePlatform": {
        "type": "string",
        "enum": [
          "reddit",
          "knowyourmeme"
        ],
        "title": "SourcePlatform",
        "description": "Supported source platforms."
      },
      "SourceStatus": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Unique source ID, e.g. reddit_memes"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Source identifier, e.g. reddit:r/memes or Reddit r/memes"
          },
          "platform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourcePlatform"
              },
              {
                "type": "string"
              }
            ],
            "title": "Platform",
            "description": "Source platform enum or string (reddit, knowyourmeme)"
          },
          "community": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community",
            "description": "Subreddit name or feed category",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Health status: ok, degraded, failing, or offline",
            "default": "ok"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count",
            "description": "Number of memes successfully cached from this source",
            "default": 0
          },
          "last_synced_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At",
            "description": "Unix timestamp of last successful sync"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error",
            "description": "Error message from last failed ingestion attempt"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Alias for last_error"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms",
            "description": "HTTP round-trip latency in milliseconds"
          }
        },
        "type": "object",
        "required": [
          "name",
          "platform"
        ],
        "title": "SourceStatus",
        "description": "Health and metric status for an individual ingestion source feed."
      }
    }
  },
  "tags": [
    {
      "name": "memes",
      "description": "Endpoints for querying newest, trending, and random memes across Reddit and Know Your Meme."
    },
    {
      "name": "sources",
      "description": "Endpoints for monitoring active ingestion sources, health status, and sync telemetry."
    },
    {
      "name": "health",
      "description": "System operational health checks, uptime metrics, and cache statistics."
    }
  ]
}